requirements.txt
Flask==3.0.3
transformers==4.46.3
torch==2.2.2
pillow==10.3.0
face-recognition==1.3.0
